backend/requirements.txt

fastapi>=0.110.0
uvicorn[standard]>=0.27.0
SQLAlchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-dotenv>=1.0.0
email-validator>=2.1.0
argon2-cffi>=23.1.0

